// web-app/src/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  graphql_public: {
    Tables: {
      [_ in never]: never
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      graphql: {
        Args: {
          extensions?: Json
          operationName?: string
          query?: string
          variables?: Json
        }
        Returns: Json
      }
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  public: {
    Tables: {
      media_items: {
        Row: {
          created_at: string
          description: string | null
          genres: string[]
          id: string
          imdb_id: string | null
          rotten_tomatoes_score: number | null
          streaming_services: Json
          title: string
          tmdb_id: string
          total_seasons: number
          type: Database["public"]["Enums"]["media_type"]
        }
        Insert: {
          created_at?: string
          description?: string | null
          genres?: string[]
          id?: string
          imdb_id?: string | null
          rotten_tomatoes_score?: number | null
          streaming_services?: Json
          title: string
          tmdb_id: string
          total_seasons?: number
          type: Database["public"]["Enums"]["media_type"]
        }
        Update: {
          created_at?: string
          description?: string | null
          genres?: string[]
          id?: string
          imdb_id?: string | null
          rotten_tomatoes_score?: number | null
          streaming_services?: Json
          title?: string
          tmdb_id?: string
          total_seasons?: number
          type?: Database["public"]["Enums"]["media_type"]
        }
        Relationships: []
      }
      profiles: {
        Row: {
          avatar_url: string | null
          created_at: string
          display_name: string
          id: string
        }
        Insert: {
          avatar_url?: string | null
          created_at?: string
          display_name: string
          id: string
        }
        Update: {
          avatar_url?: string | null
          created_at?: string
          display_name?: string
          id?: string
        }
        Relationships: []
      }
      user_media_states: {
        Row: {
          current_season: number
          id: string
          media_item_id: string
          profile_id: string
          state: Database["public"]["Enums"]["kanban_state"]
          updated_at: string
        }
        Insert: {
          current_season?: number
          id?: string
          media_item_id: string
          profile_id: string
          state?: Database["public"]["Enums"]["kanban_state"]
          updated_at?: string
        }
        Update: {
          current_season?: number
          id?: string
          media_item_id?: string
          profile_id?: string
          state?: Database["public"]["Enums"]["kanban_state"]
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "user_media_states_media_item_id_fkey"
            columns: ["media_item_id"]
            isOneToOne: false
            referencedRelation: "media_items"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "user_media_states_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      kanban_state:
        | "not_available"
        | "available"
        | "prioritised"
        | "watching"
        | "watched"
      media_type: "movie" | "tv"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never

export type CompositeTypes<
  PublicCompositeTypeNameOrOptions extends
    | keyof DefaultSchema["CompositeTypes"]
    | { schema: keyof DatabaseWithoutInternals },
  CompositeTypeName extends PublicCompositeTypeNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"]
    : never = never,
> = PublicCompositeTypeNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[PublicCompositeTypeNameOrOptions["schema"]]["CompositeTypes"][CompositeTypeName]
  : PublicCompositeTypeNameOrOptions extends keyof DefaultSchema["CompositeTypes"]
    ? DefaultSchema["CompositeTypes"][PublicCompositeTypeNameOrOptions]
    : never